Changes to cog_vi filenames and combined SelectionFunctionUtils with carpentry to make cog_vi.py.
